refactor: Rename file that ends demo survey
