+++ lib/src/routes.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:mdigit_span_tasks_ema/src/demographics/process_demographics_data.dart';
+import 'package:mdigit_span_tasks_ema/src/demographics/end_demographics_survey.dart';
 import 'package:mdigit_span_tasks_ema/src/ema/ema_screen.dart';
 import 'package:mdigit_span_tasks_ema/src/task_list/view/task_list_page.dart';
 import 'package:mdigit_span_tasks_ema/src/ui_components/loading_screen.dart';
